Move function defs from wallet api to wallet
